Add files via upload
This is a revised prototype version of the turbidity tube for strapping directly under the hull of a kayak. There are spaces for two 2cm webbing straps to pass through to hold the tube onto the kayak. There is a holder for a digital thermometer on the side. Cables from the LDR, LED and temperature sensor can pass through the cable holder on the top, and up the webbing strap to keep them safe. We suggest adding a piece of thick soft neoprene between the tube and the kayak hull to protect the sensors and wiring, and ensure a snug tight fit that will not come loose. The straps around the hull need to pass through a D-ring on the kayak for example, to make sure the whole setup doesn't simply slip off the back of the kayak. This version has not been tested on the water yet, and as such has not yet been added to the wiki instructions, but has been printed (Ender3 Pro) and tested for fit on a standard kayak.